refactor: addresses ito's comments
