+++ src/components/pages/storage/buy/StorageOffersCheckoutPage.tsx
@@ -11,7 +11,7 @@ import RoundedCard from 'components/atoms/RoundedCard'
 import { CombinedPriceCell, JobDoneBox } from 'components/molecules'
 import GridRow from 'components/molecules/storage/buy/GridRow'
 import PlanOption from 'components/molecules/storage/buy/PlanOption'
-import RifSelect from 'components/molecules/storage/buy/RifSelect'
+import RifSelect from 'components/molecules/RifSelect'
 import ConfigPurchaseCard from 'components/organisms/storage/buy/ConfigPurchaseCard'
 import PinningCard from 'components/organisms/storage/buy/PinningCard'
 import StorageOrderDescription from 'components/organisms/storage/buy/StorageOfferDescription'
